test(memory-core): migrate SessionService concurrency test to Playwright spec (#10693)
Addresses Cycle 2 review feedback by wrapping the master lock contention loop in a proper Playwright test block with expect() assertions. Isolated the worker simulation logic into a separate worker file to prevent execution during the Playwright test discovery pass.
